Create custome user with email field instead of username. Register model. Wrote skeleton of it base serializer
