+++ users_service/serializers.py
@@ -1,0 +1,11 @@
+from django.contrib.auth import get_user_model
+from rest_framework import serializers
+
+
+class UserBaseSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = get_user_model()
+        fields = (
+            "id", "first_name", "last_name", "email", "nickname"
+        )
